FreeRTOS and machine learning sample code updat

+++ SampleCode/FreeRTOS/TicklessIdle/ParTest.c
@@ -51,11 +51,11 @@ int IsDebugFifoEmpty(void)
  */
 void RTC_IRQHandler(void)
 {
-	uint32_t u32ReadRTC_Ticks;
-	S_RTC_TIME_DATA_T sReadRTC;
+    uint32_t u32ReadRTC_Ticks;
+    S_RTC_TIME_DATA_T sReadRTC;
 
     /* To check if RTC alarm interrupt occurred */
-    if(RTC_GET_ALARM_INT_FLAG(RTC) == 1)
+    if (RTC_GET_ALARM_INT_FLAG(RTC) == 1)
     {
         /* Clear RTC alarm interrupt flag */
         RTC_CLEAR_ALARM_INT_FLAG(RTC);
@@ -65,12 +65,12 @@ void RTC_IRQHandler(void)
         u32ReadRTC_Ticks = ((RTC->TIME & RTC_TIME_HZCNT_Msk) >> RTC_TIME_HZCNT_Pos);
 
         printf("RTC Alarm: %d/%02d/%02d %02d:%02d:%02d.%02d\n",
-               sReadRTC.u32Year, sReadRTC.u32Month, sReadRTC.u32Day, sReadRTC.u32Hour, sReadRTC.u32Minute, sReadRTC.u32Second, u32ReadRTC_Ticks * 100 / 128);
+               sReadRTC.u32Year, sReadRTC.u32Month, sReadRTC.u32Day, sReadRTC.u32Hour, sReadRTC.u32Minute, sReadRTC.u32Second, (int32_t)u32ReadRTC_Ticks * 100 / 128);
 
     }
 
     /* To check if RTC tick interrupt occurred */
-    if(RTC_GET_TICK_INT_FLAG(RTC) == 1)
+    if (RTC_GET_TICK_INT_FLAG(RTC) == 1)
     {
         /* Clear RTC tick interrupt flag */
         RTC_CLEAR_TICK_INT_FLAG(RTC);
@@ -88,30 +88,30 @@ void RTC_IRQHandler(void)
  */
 void GPB_IRQHandler(void)
 {
-	uint32_t u32ReadRTC_Ticks;
-	S_RTC_TIME_DATA_T sReadRTC;
+    uint32_t u32ReadRTC_Ticks;
+    S_RTC_TIME_DATA_T sReadRTC;
 
-    if(GPIO_GET_INT_FLAG(PB, BIT1))      /* To check if PB.1 interrupt occurred */
+    if (GPIO_GET_INT_FLAG(PB, BIT1))     /* To check if PB.1 interrupt occurred */
     {
         GPIO_CLR_INT_FLAG(PB, BIT1);
         printf("CPU woken up by external interrupt (PB.1 INT occurred).\n");
     }
-    else if(GPIO_GET_INT_FLAG(PB, BIT2)) /* To check if PB.2 interrupt occurred */
+    else if (GPIO_GET_INT_FLAG(PB, BIT2)) /* To check if PB.2 interrupt occurred */
     {
         GPIO_CLR_INT_FLAG(PB, BIT2);
         printf("CPU woken up by external interrupt (PB.2 INT occurred).\n");
     }
-    else if(GPIO_GET_INT_FLAG(PB, BIT3)) /* To check if PB.3 interrupt occurred */
+    else if (GPIO_GET_INT_FLAG(PB, BIT3)) /* To check if PB.3 interrupt occurred */
     {
         GPIO_CLR_INT_FLAG(PB, BIT3);
         printf("CPU woken up by external interrupt (PB.3 INT occurred).\n");
     }
-    else if(GPIO_GET_INT_FLAG(PB, BIT4)) /* To check if PB.4 interrupt occurred */
+    else if (GPIO_GET_INT_FLAG(PB, BIT4)) /* To check if PB.4 interrupt occurred */
     {
         GPIO_CLR_INT_FLAG(PB, BIT4);
         printf("CPU woken up by external interrupt (PB.4 INT occurred).\n");
     }
-    else if(GPIO_GET_INT_FLAG(PB, BIT5)) /* To check if PB.5 interrupt occurred */
+    else if (GPIO_GET_INT_FLAG(PB, BIT5)) /* To check if PB.5 interrupt occurred */
     {
         GPIO_CLR_INT_FLAG(PB, BIT5);
         printf("CPU woken up by external interrupt (PB.5 INT occurred).\n");
@@ -128,13 +128,13 @@ void GPB_IRQHandler(void)
     u32ReadRTC_Ticks = ((RTC->TIME & RTC_TIME_HZCNT_Msk) >> RTC_TIME_HZCNT_Pos);
 
     printf("IO wakup up time: %d/%02d/%02d %02d:%02d:%02d.%02d\n",
-            sReadRTC.u32Year, sReadRTC.u32Month, sReadRTC.u32Day, sReadRTC.u32Hour, sReadRTC.u32Minute, sReadRTC.u32Second, u32ReadRTC_Ticks * 100 / 128);
+           sReadRTC.u32Year, sReadRTC.u32Month, sReadRTC.u32Day, sReadRTC.u32Hour, sReadRTC.u32Minute, sReadRTC.u32Second, (int32_t)u32ReadRTC_Ticks * 100 / 128);
 }
 
 void PowerDownFunction(void)
 {
 #if defined(PMC_IDLE_MODE)
-	PMC_Idle();
+    PMC_Idle();
 #else
 
     uint32_t u32TimeOutCnt;
@@ -142,8 +142,9 @@ void PowerDownFunction(void)
     PMC_SetPowerDownMode(PMC_NPD0, PMC_PLCTL_PLSEL_PL1);
     /* To check if all the debug messages are finished */
     u32TimeOutCnt = SystemCoreClock; /* 1 second time-out */
-    while(IsDebugFifoEmpty() == 0)
-		if(--u32TimeOutCnt == 0) break;
+
+    while (IsDebugFifoEmpty() == 0)
+        if (--u32TimeOutCnt == 0) break;
 
     /* Enter to Power-down mode */
     PMC_PowerDown();
@@ -153,7 +154,7 @@ void PowerDownFunction(void)
 void vParTestInitialise(void)
 {
     uint32_t u32Pin;
-	S_RTC_TIME_DATA_T sWriteRTC;
+    S_RTC_TIME_DATA_T sWriteRTC;
 
     SYS_UnlockReg();
 
@@ -166,7 +167,7 @@ void vParTestInitialise(void)
     /* Enable External LXT clock */
     CLK_EnableXtalRC(CLK_SRCCTL_LXTEN_Msk);
 
-    /* Enable HXT clock */	
+    /* Enable HXT clock */
     CLK_EnableXtalRC(CLK_SRCCTL_HXTEN_Msk);
     /* Waiting for HXT clock ready */
     CLK_WaitClockReady(CLK_STATUS_HXTSTB_Msk);
@@ -176,10 +177,10 @@ void vParTestInitialise(void)
     /* Waiting for LXT clock ready */
     CLK_WaitClockReady(CLK_STATUS_LXTSTB_Msk);
 
-    /* Switch SCLK clock source to APLL0 and Enable APLL0 180MHz clock */    
+    /* Switch SCLK clock source to APLL0 and Enable APLL0 180MHz clock */
     CLK_SetBusClock(CLK_SCLKSEL_SCLKSEL_APLL0, FREQ_180MHZ);
 
-	/* Update System Core Clock */
+    /* Update System Core Clock */
     /* User can use SystemCoreClockUpdate() to calculate SystemCoreClock. */
     SystemCoreClockUpdate();
 
@@ -190,7 +191,7 @@ void vParTestInitialise(void)
     /* Enable UART module clock */
     CLK_EnableModuleClock(UART0_MODULE);
 
-	/* Select UART0 module clock source as HIRC and UART0 module clock divider as 1 */
+    /* Select UART0 module clock source as HIRC and UART0 module clock divider as 1 */
     CLK_SetModuleClock(UART0_MODULE, CLK_UARTSEL0_UART0SEL_HXT, CLK_UARTDIV0_UART0DIV(1));
 
     /*---------------------------------------------------------------------------------------------------------*/
@@ -229,7 +230,7 @@ void vParTestInitialise(void)
     /* Enable RTC module clock */
     CLK_EnableModuleClock(RTC0_MODULE);
 
-	/* Set LXT as RTC clock source */
+    /* Set LXT as RTC clock source */
     RTC_SetClockSource(RTC_CLOCK_SOURCE_LXT);
 
     /* Enable RTC NVIC */
@@ -245,14 +246,15 @@ void vParTestInitialise(void)
     sWriteRTC.u32DayOfWeek  = RTC_FRIDAY;
     sWriteRTC.u32TimeScale  = RTC_CLOCK_24;
 
-    if(RTC_Open(&sWriteRTC) != 0)
+    if (RTC_Open(&sWriteRTC) != 0)
     {
         printf("\n RTC initial fail!!");
         printf("\n Please check h/w setting!!");
-        while(1);
+
+        while (1);
     }
 
-	RTC_SetTickPeriod(RTC_TICK_1_128_SEC);
+    RTC_SetTickPeriod(RTC_TICK_1_128_SEC);
 
     /* Enable sub-second counter */
     RTC->CLKFMT |= RTC_CLKFMT_HZCNTEN_Msk;
@@ -260,8 +262,10 @@ void vParTestInitialise(void)
 
     /* Configure PB.1~5 as Input mode and enable interrupt by rising or falling edge trigger */
     GPIO_SetMode(PB, (BIT1 | BIT2 | BIT3 | BIT4 | BIT5), GPIO_MODE_INPUT);
-    for(u32Pin = 1; u32Pin < 6; u32Pin++)
+
+    for (u32Pin = 1; u32Pin < 6; u32Pin++)
         GPIO_EnableInt(PB, u32Pin, GPIO_INT_BOTH_EDGE);
+
     NVIC_EnableIRQ(GPB_IRQn);
 
     /* LED control */
