+++ SampleCode/FreeRTOS/TrustZone/tz_demo/tz_demo.c
@@ -32,7 +32,7 @@
 /* Non-Secure callable functions. */
 #include "nsc_functions.h"
 
-#define NUM_TASKS	4
+#define NUM_TASKS   4
 
 /**
  * @brief Counter incremented in the callback which is called from the secure
@@ -44,20 +44,20 @@
  * 4 bytes and upto 32 bytes will also fall in the same MPU region and the task
  * having access to ulNonSecureCounter will also have access to all those items.
  */
-static uint32_t ulNonSecureCounter[ NUM_TASKS ] __attribute__( ( aligned( 32 ) ) ) = { 0 };
+static uint32_t ulNonSecureCounter[ NUM_TASKS ] __attribute__((aligned(32))) = { 0 };
 /*-----------------------------------------------------------*/
 
 /**
  * @brief Creates all the tasks for TZ demo.
  */
-void vStartTZDemo( void );
+void vStartTZDemo(void);
 
 /**
  * @brief Increments the ulNonSecureCounter.
  *
  * This function is called from the secure side.
  */
-static void prvCallback( void *pvParameters );
+static void prvCallback(void *pvParameters);
 
 /**
  * @brief Implements the task which calls the functions exported from the secure
@@ -65,27 +65,27 @@ static void prvCallback( void *pvParameters );
  *
  * @param pvParameters[in] Parameters as passed during task creation.
  */
-static void prvSecureCallingTask( void * pvParameters );
+static void prvSecureCallingTask(void *pvParameters);
 /*-----------------------------------------------------------*/
 
-void vStartTZDemo( void )
+void vStartTZDemo(void)
 {
-	int i;
-	
-	for( i = 0 ; i < NUM_TASKS; i ++)
-		xTaskCreate( prvSecureCallingTask,
-				"SecCalling",
-				configMINIMAL_STACK_SIZE,
-				(void *) i,
-				tskIDLE_PRIORITY,
-				NULL ); /* FREERTOS_SYSTEM_CALL */
+    int i;
+
+    for (i = 0 ; i < NUM_TASKS; i ++)
+        xTaskCreate(prvSecureCallingTask,
+                    "SecCalling",
+                    configMINIMAL_STACK_SIZE,
+                    (void *) i,
+                    tskIDLE_PRIORITY,
+                    NULL);  /* FREERTOS_SYSTEM_CALL */
 
 }
 /*-----------------------------------------------------------*/
 
-static void prvCallback( void *pvParameters)
+static void prvCallback(void *pvParameters)
 {
-	int32_t i32TaskNo = (int32_t)pvParameters;
+    int32_t i32TaskNo = (int32_t)pvParameters;
 
     /* This function is called from the secure side. Just increment the counter
      * here. The check that this counter keeps incrementing is performed in the
@@ -94,22 +94,22 @@ static void prvCallback( void *pvParameters)
 }
 /*-----------------------------------------------------------*/
 
-static void prvSecureCallingTask( void * pvParameters )
+static void prvSecureCallingTask(void *pvParameters)
 {
-	char szInfo[100];
+    char szInfo[100];
     uint32_t ulLastSecureCounter = 0, ulLastNonSecureCounter = 0;
     uint32_t ulCurrentSecureCounter = 0;
-	int32_t i32TaskNo = (int32_t)pvParameters;
-	S_NSCFUN_PARAMETER sFunParam;
-	
-	sFunParam.i32TaskNo = i32TaskNo;
-	sFunParam.ulSecureCounter = 0;
+    int32_t i32TaskNo = (int32_t)pvParameters;
+    S_NSCFUN_PARAMETER sFunParam;
+
+    sFunParam.i32TaskNo = i32TaskNo;
+    sFunParam.ulSecureCounter = 0;
 
     /* This task calls secure side functions. So allocate a secure context for
      * it. */
-    portALLOCATE_SECURE_CONTEXT( configMINIMAL_SECURE_STACK_SIZE );
+    portALLOCATE_SECURE_CONTEXT(configMINIMAL_SECURE_STACK_SIZE);
 
-    for( ; ; )
+    for (; ;)
     {
         /* Call the secure side function. It does two things:
          * - It calls the supplied function (prvCallback) which in turn
@@ -118,24 +118,24 @@ static void prvSecureCallingTask( void * pvParameters )
          * Therefore at the end of this function call both the secure and
          * non-secure counters must have been incremented.
          */
-        ulCurrentSecureCounter = NSCFunction( prvCallback, &sFunParam);
+        ulCurrentSecureCounter = NSCFunction(prvCallback, &sFunParam);
 
         /* Make sure that both the counters are incremented. */
-        configASSERT( ulCurrentSecureCounter == ulLastSecureCounter + 1 );
-        configASSERT( ulNonSecureCounter[ i32TaskNo ] == ulLastNonSecureCounter + 1 );
+        configASSERT(ulCurrentSecureCounter == ulLastSecureCounter + 1);
+        configASSERT(ulNonSecureCounter[ i32TaskNo ] == ulLastNonSecureCounter + 1);
 
-		sprintf(szInfo, "ulCurrentSecureCounter %d \n", ulCurrentSecureCounter);
-		NSCPrintf(szInfo);
-		sprintf(szInfo, "ulNonSecureCounter[ %d ] %d \n", i32TaskNo, ulNonSecureCounter[ i32TaskNo ]);
-		NSCPrintf(szInfo);
-		
+        sprintf(szInfo, "ulCurrentSecureCounter %d \n", (int32_t)ulCurrentSecureCounter);
+        NSCPrintf(szInfo);
+        sprintf(szInfo, "ulNonSecureCounter[ %d ] %d \n", i32TaskNo, ulNonSecureCounter[ i32TaskNo ]);
+        NSCPrintf(szInfo);
+
         /* Update the last values for both the counters. */
         ulLastSecureCounter = ulCurrentSecureCounter;
         ulLastNonSecureCounter = ulNonSecureCounter[ i32TaskNo ];
 
         /* Wait for a second. */
-//        vTaskDelay( pdMS_TO_TICKS( 1000 ) );
-        vTaskDelay( pdMS_TO_TICKS( 10 ) );
+        //        vTaskDelay( pdMS_TO_TICKS( 1000 ) );
+        vTaskDelay(pdMS_TO_TICKS(10));
     }
 }
 /*-----------------------------------------------------------*/
